+++ pkg/config/config.go
@@ -104,9 +104,10 @@ type Versioner struct {
 
 // Language represents Compute language specific configuration.
 type Language struct {
-	CPP  CPP  `toml:"cpp"`
-	Go   Go   `toml:"go"`
-	Rust Rust `toml:"rust"`
+	CPP    CPP    `toml:"cpp"`
+	Go     Go     `toml:"go"`
+	Python Python `toml:"python"`
+	Rust   Rust   `toml:"rust"`
 }
 
 // Go represents Go Compute language specific configuration.
@@ -148,6 +149,15 @@ type CPP struct {
 	WasmWasiTarget string `toml:"wasm_wasi_target"`
 }
 
+// Python represents Python Compute language specific configuration.
+type Python struct {
+	// ToolchainConstraint is the Python version that we support (host Python).
+	ToolchainConstraint string `toml:"toolchain_constraint"`
+
+	// UVConstraint is the UV package manager version (optional, for warnings).
+	UVConstraint string `toml:"uv_constraint"`
+}
+
 // Profiles represents multiple profile accounts.
 type Profiles map[string]*Profile
 
@@ -183,6 +193,7 @@ type StarterKitLanguages struct {
 	Go         []StarterKit `toml:"go"`
 	JavaScript []StarterKit `toml:"javascript"`
 	Rust       []StarterKit `toml:"rust"`
+	Python     []StarterKit `toml:"python"`
 }
 
 // StarterKit represents starter kit specific configuration.